pyproject.toml
[project]
name = "rohit-samaj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "babel==2.16.0",
    "django==5.1.3",
    "django-allauth[socialaccount]>=65.5.0",
    "django-appconf==1.0.6",
    "django-browser-reload==1.17.0",
    "django-compressor==4.5.1",
    "django-cotton==1.3.0",
    "django-dotenv==1.4.2",
    "django-money==3.5.3",
    "django-stubs[compatible-mypy]>=5.1.3",
    "psycopg2-binary==2.9.10",
    "py-moneyed==3.0",
    "rcssmin==1.1.2",
    "rjsmin==1.2.2",
    "setuptools==75.6.0",
    "sqlparse==0.5.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "rohit_samaj.settings"
